ConPTY: allow for unbroken copying of long lines
On Windows, ConPTY does not support 'WRAPLINE', meaning that the lines in a copied
selection were incorrectly 'hard wrapped' with a newline character, even
if the line should not have been broken.

This change makes it such that lines are hard-wrapped if:

    * the length of the selection is >= the number of columns in the
      terminal; and

    * the final column is a space (not final letter, final column)

That is, if you make a selection that is wide as the terimal *and* the
last line *is not* a space, your selection is not broken with newlines.